+++ src/app/ps/[slug]/page.tsx
@@ -1,6 +1,6 @@
 import { psPosts } from "#posts";
 import { notFound } from "next/navigation";
-import { MDXContent } from "#mdx-content"
+import { MDXContent } from "@/components/mdx-content"
 
 export type ParamsType = Promise<{ slug: string }>;
