+++ tsconfig.json
@@ -21,7 +21,6 @@
     "paths": {
       "@/*": ["./src/*"],
       "#posts": ["./.velite"],
-      "#mdx-content": ["./src/app/blog/[slug]/components/mdx-content.tsx"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
